// src/model/Comment.ts
/*!
 * Copyright (c) 2025 Carlos Coloma Escribano
 *
 * This source code is available under a dual licensing model:
 *
 * 1. LICENSE.txt — Public license: permits inspection and internal evaluation only.
 * 2. LICENSE-CHOICE.md — Commercial license: required for any production or commercial use.
 *
 */

export interface Comment {
  id: string;
  body: string;
  created: number;
  creator: {
    name: string;
    email: string;
    support: boolean;
  };
  [key: string]: unknown;
}